refactor(nx): consolidate all dynamic config
